+++ src/command_modules/azure-cli-container/HISTORY.rst
@@ -3,6 +3,11 @@
 Release History
 ===============
 
+0.2.2
++++++
+* Remove the requirement for username and password for non dockerhub registries
+* Fix error when creating container groups from yaml file
+
 0.2.1
 +++++
 * Update PyYAML dependency to 4.2b4